pyproject.toml
[project]
name = "bdp-oct25-exam-bdp-oct25-group7"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.13.3",
    "dotenv>=0.9.9",
    "fastapi>=0.128.0",
    "httpx>=0.28.1",
    "imblearn>=0.0",
    "openpyxl>=3.1.5",
    "pandas>=2.0.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
    "streamlit>=1.42.0",
    "streamlit-code-editor>=0.1.0",
    "uvicorn>=0.40.0",
    "xgboost>=3.2.0",
]

[tool.setuptools.packages.find]
include = ["code*"]
